Changed gbifId from int to long

+++ sdks/keygen/src/main/java/org/gbif/pipelines/keygen/api/KeyLookupResult.java
@@ -13,7 +13,7 @@ import lombok.Getter;
 @EqualsAndHashCode
 public class KeyLookupResult {
 
-  private final int key;
+  private final long key;
   private final boolean created;
 
 }
